[C] 3.1 Parsing Logic for variables and identifiers
- incomplete binding_def.rs
- project structure shift to modular
- Test passes
=======[20250618]=======
